requirements.txt
python-telegram-bot==20.0
rembg[no-numpy]>=2.0.41
Pillow
onnxruntime